Fix
Functions shouldn't be called when set to when_pressed, that'll run the function once (immediately) and send the return value (None) into when_pressed. Instead when_pressed should be set to a function reference or lambda. I've removed the parentheses to make it a function reference.

i.e. "run this function when the button is pressed"